+++ galaxy_ml/model_validations/_fit_and_score.py
@@ -1,0 +1,233 @@
+import numpy as np
+import numbers
+import time
+import warnings
+
+from sklearn.exceptions import FitFailedWarning
+from sklearn.model_selection._validation import _index_param_value
+from sklearn.model_selection._validation import _score
+from sklearn.utils.metaestimators import _safe_split
+from sklearn.utils.validation import _num_samples
+# from sklearn.utils import _message_with_time
+from traceback import format_exception_only
+
+
+def _fit_and_score(estimator, X, y, scorer, train, test, verbose,
+                   parameters, fit_params, return_train_score=False,
+                   return_parameters=False, return_n_test_samples=False,
+                   return_times=False, return_estimator=False,
+                   error_score=np.nan):
+    """override the sklearn.model_selection._validation._fit_and_score
+
+    Parameters
+    ----------
+    estimator : estimator object implementing 'fit'
+        The object to use to fit the data.
+    X : array-like of shape at least 2D
+        The data to fit.
+    y : array-like, optional, default: None
+        The target variable to try to predict in the case of
+        supervised learning.
+    scorer : A single callable or dict mapping scorer name to the callable
+        If it is a single callable, the return value for ``train_scores`` and
+        ``test_scores`` is a single float.
+        For a dict, it should be one mapping the scorer name to the scorer
+        callable object / function.
+        The callable object / fn should have signature
+        ``scorer(estimator, X, y)``.
+    train : array-like, shape (n_train_samples,)
+        Indices of training samples.
+    test : array-like, shape (n_test_samples,)
+        Indices of test samples.
+    verbose : integer
+        The verbosity level.
+    error_score : 'raise' or numeric
+        Value to assign to the score if an error occurs in estimator fitting.
+        If set to 'raise', the error is raised.
+        If a numeric value is given, FitFailedWarning is raised. This parameter
+        does not affect the refit step, which will always raise the error.
+    parameters : dict or None
+        Parameters to be set on the estimator.
+    fit_params : dict or None
+        Parameters that will be passed to ``estimator.fit``.
+    return_train_score : boolean, optional, default: False
+        Compute and return score on training set.
+    return_parameters : boolean, optional, default: False
+        Return parameters that has been used for the estimator.
+    return_n_test_samples : boolean, optional, default: False
+        Whether to return the ``n_test_samples``
+    return_times : boolean, optional, default: False
+        Whether to return the fit/score times.
+    return_estimator : boolean, optional, default: False
+        Whether to return the fitted estimator.
+    Returns
+    -------
+    train_scores : dict of scorer name -> float, optional
+        Score on training set (for all the scorers),
+        returned only if `return_train_score` is `True`.
+    test_scores : dict of scorer name -> float, optional
+        Score on testing set (for all the scorers).
+    n_test_samples : int
+        Number of test samples.
+    fit_time : float
+        Time spent for fitting in seconds.
+    score_time : float
+        Time spent for scoring in seconds.
+    parameters : dict or None, optional
+        The parameters that have been evaluated.
+    estimator : estimator object
+        The fitted estimator
+    """
+    if verbose > 1:
+        if parameters is None:
+            msg = ''
+        else:
+            msg = '%s' % (', '.join('%s=%s' % (k, v)
+                          for k, v in parameters.items()))
+        print("[CV] %s %s" % (msg, (64 - len(msg)) * '.'))
+
+    # Adjust length of sample weights
+    fit_params = fit_params if fit_params is not None else {}
+    fit_params = {k: _index_param_value(X, v, train)
+                  for k, v in fit_params.items()}
+
+    train_scores = {}
+    if parameters is not None:
+        estimator.set_params(**parameters)
+
+    start_time = time.time()
+
+    X_train, y_train = _safe_split(estimator, X, y, train)
+    X_test, y_test = _safe_split(estimator, X, y, test, train)
+
+    is_multimetric = not callable(scorer)
+    n_scorers = len(scorer.keys()) if is_multimetric else 1
+
+    ##########################################################
+    # Changes on sklearn.model_selection._search
+    # for param in estimator.get_params().keys():
+    #    # suppose work for both pipeline or single
+    #    # keras model
+    #    if param.endswith('validation_data'):
+    #        fit_params.update(
+    #            {param: (X_test, y_test)})
+    #        break
+    # Changes end
+    ##########################################################
+    try:
+        if y_train is None:
+            estimator.fit(X_train, **fit_params)
+        else:
+            estimator.fit(X_train, y_train, **fit_params)
+
+    except Exception as e:
+        # Note fit time as time until error
+        fit_time = time.time() - start_time
+        score_time = 0.0
+        if error_score == 'raise':
+            raise
+        elif error_score == 'raise-deprecating':
+            warnings.warn("From version 0.22, errors during fit will result "
+                          "in a cross validation score of NaN by default. Use "
+                          "error_score='raise' if you want an exception "
+                          "raised or error_score=np.nan to adopt the "
+                          "behavior from version 0.22.",
+                          FutureWarning)
+            raise
+        elif isinstance(error_score, numbers.Number):
+            if is_multimetric:
+                test_scores = dict(zip(scorer.keys(),
+                                   [error_score, ] * n_scorers))
+                if return_train_score:
+                    train_scores = dict(zip(scorer.keys(),
+                                        [error_score, ] * n_scorers))
+            else:
+                test_scores = error_score
+                if return_train_score:
+                    train_scores = error_score
+            warnings.warn("Estimator fit failed. The score on this train-test"
+                          " partition for these parameters will be set to %f. "
+                          "Details: \n%s" %
+                          (error_score, format_exception_only(type(e), e)[0]),
+                          FitFailedWarning)
+        else:
+            raise ValueError("error_score must be the string 'raise' or a"
+                             " numeric value. (Hint: if using 'raise', please"
+                             " make sure that it has been spelled correctly.)")
+
+    else:
+        fit_time = time.time() - start_time
+
+        #######################################################################
+        #######################################################################
+        if estimator.__class__.__name__ == 'KerasGBatchClassifier':
+            test_scores = estimator.evaluate(X_test, y_test,
+                                             scorer, is_multimetric)
+        else:
+            # _score will return dict if is_multimetric is True
+            test_scores = _score(estimator, X_test, y_test,
+                                 scorer, is_multimetric)
+        score_time = time.time() - start_time - fit_time
+        if return_train_score:
+            if estimator.__class__.__name__ == 'KerasGBatchClassifier':
+                train_scores = estimator.evaluate(X_train, y_train,
+                                                  scorer, is_multimetric)
+            else:
+                train_scores = _score(estimator, X_train, y_train, scorer,
+                                      is_multimetric)
+        #######################################################################
+        #######################################################################
+    if verbose > 2:
+        if is_multimetric:
+            for scorer_name in sorted(test_scores):
+                msg += ", %s=" % scorer_name
+                if return_train_score:
+                    msg += "(train=%.3f," % train_scores[scorer_name]
+                    msg += " test=%.3f)" % test_scores[scorer_name]
+                else:
+                    msg += "%.3f" % test_scores[scorer_name]
+        else:
+            msg += ", score="
+            msg += ("%.3f" % test_scores if not return_train_score else
+                    "(train=%.3f, test=%.3f)" % (train_scores, test_scores))
+
+    if verbose > 1:
+        total_time = score_time + fit_time
+        print(_message_with_time('CV', msg, total_time))
+
+    ret = [train_scores, test_scores] if return_train_score else [test_scores]
+
+    if return_n_test_samples:
+        ret.append(_num_samples(X_test))
+    if return_times:
+        ret.extend([fit_time, score_time])
+    if return_parameters:
+        ret.append(parameters)
+    if return_estimator:
+        ret.append(estimator)
+    return ret
+
+
+# `sklearn.utils._message_with_time` is available from v0.21.x.
+def _message_with_time(source, message, time):
+    """Create one line message for logging purposes
+    Parameters
+    ----------
+    source : str
+        String indicating the source or the reference of the message
+    message : str
+        Short message
+    time : int
+        Time in seconds
+    """
+    start_message = "[%s] " % source
+
+    # adapted from joblib.logger.short_format_time without the Windows -.1s
+    # adjustment
+    if time > 60:
+        time_str = "%4.1fmin" % (time / 60)
+    else:
+        time_str = " %5.1fs" % time
+    end_message = " %s, total=%s" % (message, time_str)
+    dots_len = (70 - len(start_message) - len(end_message))
+    return "%s%s%s" % (start_message, dots_len * '.', end_message)
